Fix for security review

+++ cie-pkcs11/PKCS11/Mechanism.cpp
@@ -581,7 +581,7 @@ namespace p11 {
 
 		// se non posso levare il padding, la firma ha
 		// qualcosa di sbagliato
-		DWORD dwPadLen;
+		unsigned long dwPadLen;
 		try {
 			dwPadLen = RemovePaddingBT1(baPlainSignature);
 		}
@@ -709,7 +709,7 @@ namespace p11 {
 
 			// devo levare il padding BT2 da paddedData e copiare tutto in unpaddedData;
 			// se non posso levare il padding i dati criptati non sono validi!
-			DWORD dwPaddingLen;
+			unsigned long dwPaddingLen;
 		try {
 			dwPaddingLen = RemovePaddingBT2(paddedData);
 		}
